Moving ExtractArchive and JUnRarTestUtil to unrar module
